apply a lot of very questionable fixes
